+++ custom_components/fermax_blue/button.py
@@ -27,10 +27,7 @@ async def async_setup_entry(
 
     for coordinator in coordinators:
         for door_name, door in coordinator.pairing.access_doors.items():
-            if door.visible:
-                entities.append(
-                    FermaxOpenDoorButton(coordinator, door_name, door.title)
-                )
+            entities.append(FermaxOpenDoorButton(coordinator, door_name, door.title))
         entities.append(FermaxCameraPreviewButton(coordinator))
         entities.append(FermaxF1Button(coordinator))
         entities.append(FermaxCallGuardButton(coordinator))